created a new component for individual groups
